# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api",
    "paths": {
        "/address/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Create new address",
                "parameters": [
                    {
                        "description": "address data",
                        "name": "address_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.NewAddressDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AddressRespOKDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ValidationRespErrDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
        "/address/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Delete address",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3afc3021-9395-11ed-a8b6-d8bbc1a27045",
                        "description": "address  uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.RespOKDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
        "/address/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all addresses from auth user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Get auth user addresses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AddressesRespOKDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
        "/address/update/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Update address",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3afc3021-9395-11ed-a8b6-d8bbc1a27045",
                        "description": "address  uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "address data",
                        "name": "address_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.UpdateAddressDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AddressRespOKDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.ValidationRespErrDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.RespErrDTO"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ZaphCode_clean-arch_src_api_dtos.AddressDTO": {
            "type": "object",
            "required": [
                "city",
                "country",
                "line1",
                "name",
                "postal_code",
                "state"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "Los Angeles"
                },
                "country": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "USA"
                },
                "created_at": {
                    "type": "integer",
                    "example": 1674405183
                },
                "id": {
                    "type": "string",
                    "example": "8ded83fe-93c8-11ed-ab0f-d8bbc1a27048"
                },
                "line1": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "Lolipop"
                },
                "line2": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "Wolfstreet"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "main card"
                },
                "postal_code": {
                    "type": "string",
                    "example": "24156"
                },
                "state": {
                    "type": "string",
                    "maxLength": 40,
                    "example": "California"
                },
                "updated_at": {
                    "type": "integer",
                    "example": 1674405181
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.AddressRespOKDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AddressDTO"
                },
                "message": {
                    "type": "string",
                    "example": "Data retrived!"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.AddressesRespOKDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ZaphCode_clean-arch_src_api_dtos.AddressDTO"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Data retrived!"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.AuthRespErrDTO": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "the token is expired"
                },
                "message": {
                    "type": "string",
                    "example": "invalid access token"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.DetailRespErrDTO": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "error querying the database"
                },
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.NewAddressDTO": {
            "type": "object",
            "required": [
                "city",
                "country",
                "line1",
                "name",
                "postal_code",
                "state"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "Los Angeles"
                },
                "country": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "USA"
                },
                "line1": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "Lolipop"
                },
                "line2": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "Wolfstreet"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "main card"
                },
                "postal_code": {
                    "type": "string",
                    "example": "24156"
                },
                "state": {
                    "type": "string",
                    "maxLength": 40,
                    "example": "California"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.RespErrDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
        "github_com_ZaphCode_clean-arch_src_api_dtos.RespOKDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Data retrived!"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.UpdateAddressDTO": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "Los Angeles"
                },
                "country": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "USA"
                },
                "line1": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "Lolipop"
                },
                "line2": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "Wolfstreet"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "main card"
                },
                "postal_code": {
                    "type": "string",
                    "example": "24156"
                },
                "state": {
                    "type": "string",
                    "maxLength": 40,
                    "example": "California"
                }
            }
        },
...
        "github_com_ZaphCode_clean-arch_src_api_dtos.ValidationRespErrDTO": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "field": {
                                "type": "string",
                                "example": "Email"
                            },
                            "message": {
                                "type": "string",
                                "example": "invalid email"
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "one or more field are invalid"
                },
                "status": {
                    "type": "string",
                    "example": "failure"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "X-Access-Token",
            "in": "header"
        }
    }
}